Add simple heightmap based terrain and move voxel workouts to a standalone crate
