+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dotrix"
-version = "0.3.0"
+version = "0.4.0"
 authors = [
   "Elias Kartashov <elias@lowenware.com>",
   "Štěpán Wünsch <sw@lowenware.com>",
@@ -43,7 +43,7 @@ path = "./dotrix_egui"
 optional = true
 
 [dependencies.dotrix_terrain]
-version = "0.3"
+version = "0.4"
 path = "./dotrix_terrain"
 optional = true
