+++ dotrix_voxel/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dotrix_terrain"
-version = "0.4.0"
+version = "0.3.0"
 authors = ["Elias Kartashov <elias@lowenware.com>"]
 edition = "2018"
